docs/AI/GitHub_Copilot_SDK/examples/requirements.txt
github-copilot-sdk==0.1.10
python-dotenv
